Split angular correction factor into horizontal and vertical components
Split the angular correction impulse in BogieComponent.Simulate() onto 'horizontal' component (rotation axis along the vertical) and the reminder. The horizontal component is now regulated by a separate field OrientationConvergenceFactorHorizontal. When its value equals that of OrientationConvergenceFactor the algorithm is equivalent to its previous form.
Added a field 'OrientationConvergenceFactorHorizontal' to the BogieComponentDefinition (and its object builder) with the default value set to equal that of 'OrientationConvergenceFactor'.

Code cleanup